feat: add session-scoped prompt suppression and split provider helpers
**✨ Features**
- Keep the review prompt suppressed for the rest of the active VS Code session after “Maybe Later” or dismissal. The prompt eligibility resets on restart by tracking `vscode.env.sessionId` in memory only.
- Preserve call-time routing by resolving config from `options.configuration` when valid, otherwise falling back to the in-memory registry. This also merges workspace toggles onto both paths.
- Tighten parameter filtering so unsupported OpenAI params are stripped using model capability data, with static fallbacks for known model families.
- Redact quota failures only when they come from known tool results, while logging low-confidence quota phrases without redaction.

**🧪 Tests**
- Expand review prompt tests to cover session suppression, dismissal behavior, and re-prompting after a session change.
- Add coverage for the extracted config, parameter filtering, and quota redaction logic through the provider base seams.

**🧹 Chores**
- Extract call-time config resolution, parameter filtering, and quota redaction into standalone helper modules to keep the provider base thinner and easier to test.
- Update VS Code engine guidance in `AGENTS.md` to reflect the newer `^1.125.0` target.
